refactor(emit): shard object spread helpers
