+++ crates/tsz-emitter/src/emitter/module_emission/mod.rs
@@ -1,3 +1,5 @@
 mod core;
 mod exports;
 mod imports;
+#[cfg(test)]
+mod object_spread_tests;
